style: 서비스 테스트 분리
